Bluetooth: Mesh: Remove `bt_mesh_net_start`
Move `bt_mesh_net_start` functionality to `bt_mesh_start`.
