+++ subsys/bluetooth/mesh/main.c
@@ -24,6 +24,7 @@
 #include "adv.h"
 #include "prov.h"
 #include "net.h"
+#include "subnet.h"
 #include "app_keys.h"
 #include "rpl.h"
 #include "beacon.h"
@@ -344,7 +345,35 @@ static void model_start(struct bt_mesh_model *mod, struct bt_mesh_elem *elem,
 
 int bt_mesh_start(void)
 {
-	bt_mesh_net_start();
+	if (bt_mesh_beacon_get() == BT_MESH_BEACON_ENABLED) {
+		bt_mesh_beacon_enable();
+	} else {
+		bt_mesh_beacon_disable();
+	}
+
+	if (IS_ENABLED(CONFIG_BT_MESH_GATT_PROXY) &&
+	    bt_mesh_gatt_proxy_get() != BT_MESH_GATT_PROXY_NOT_SUPPORTED) {
+		bt_mesh_proxy_gatt_enable();
+		bt_mesh_adv_update();
+	}
+
+	if (IS_ENABLED(CONFIG_BT_MESH_LOW_POWER)) {
+		bt_mesh_lpn_init();
+	} else {
+		bt_mesh_scan_enable();
+	}
+
+	if (IS_ENABLED(CONFIG_BT_MESH_FRIEND)) {
+		bt_mesh_friend_init();
+	}
+
+	if (IS_ENABLED(CONFIG_BT_MESH_PROV)) {
+		struct bt_mesh_subnet *sub = bt_mesh_subnet_next(NULL);
+		uint16_t addr = bt_mesh_primary_addr();
+
+		bt_mesh_prov_complete(sub->net_idx, addr);
+	}
+
 	bt_mesh_model_foreach(model_start, NULL);
 
 	return 0;
